scraper/requirements.txt
# Existing dependencies (maintained for compatibility)
gspread==5.12.4
oauth2client==4.1.3
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.2

# Pydantic AI and validation dependencies
pydantic==2.5.3
pydantic-ai==0.0.12
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Date parsing utilities
python-dateutil==2.8.2

# HTTP requests with retry functionality
urllib3==2.2.0

# XML parsing for custom scrapers
lxml==4.9.4

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Code quality and type checking
ruff==0.1.9
mypy==1.8.0

# Async support
asyncio==3.4.3

# Additional utilities for content processing
hashlib-compat==1.0.1